requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
djangorestframework==3.14.0
Pillow==9.4.0
pip==21.3.1
pytz==2022.7
setuptools==58.5.3
sqlparse==0.4.3
tzdata==2022.7
wheel==0.37.0
